+++ src/init-action.ts
@@ -229,18 +229,14 @@ async function run() {
   core.exportVariable(EnvVar.JOB_RUN_UUID, uuidV4());
 
   try {
-    if (
-      !(await sendStatusReport(
-        await createStatusReportBase(
-          "init",
-          "starting",
-          startedAt,
-          await checkDiskUsage(logger),
-        ),
-      ))
-    ) {
-      return;
-    }
+    await sendStatusReport(
+      await createStatusReportBase(
+        "init",
+        "starting",
+        startedAt,
+        await checkDiskUsage(logger),
+      ),
+    );
 
     const codeQLDefaultVersionInfo = await features.getDefaultCliVersion(
       gitHubVersion.type,